google-shopping-sync/requirements.txt
gkeepapi>=0.16.0
python-dotenv
requests
